Enhance CameraLocationMarkers component to utilize camera status data
This commit updates the CameraLocationMarkers component to accept camera status data from the API. It modifies the marker attributes to remove color dependency on zoom level and introduces a new method to determine marker color based on camera status. The Map component is also updated to pass the retrieved camera data to CameraLocationMarkers, improving the visual representation of camera statuses on the map.